+++ commands/LimitBuyCommand.go
@@ -90,13 +90,12 @@ func (base *LimitBuyCommand) Prepare() error {
 	return nil
 }
 
+// TODO: consolidate execute across commands
 func (base LimitBuyCommand) Execute() error {
 	if v := reflect.ValueOf(base.Opts); v.IsZero() {
 		return errors.New("please call Prepare()")
 	}
-	// place order
-	// use ask price in quote to buy or sell
-	// time in force defaults to "good till canceled(gtc)"
+
 	orderRes, orderErr := base.RhClient.MakeOrder(base.Ins, base.Opts)
 
 	if orderErr != nil {