+++ commands/MarketSellCommand.go
@@ -2,6 +2,7 @@ package commands
 
 import (
 	"errors"
+	log "github.com/sirupsen/logrus"
 	"reflect"
 	"strings"
 
@@ -58,7 +59,18 @@ func (base *MarketSellCommand) Prepare() error {
 	return nil
 }
 
-func (base MarketSellCommand) Execute() (err error) {
-	_, err = base.RhClient.MakeOrder(base.Ins, base.Opts)
-	return
+func (base MarketSellCommand) Execute() error {
+	if v := reflect.ValueOf(base.Opts); v.IsZero() {
+		return errors.New("please call Prepare()")
+	}
+
+	orderRes, orderErr := base.RhClient.MakeOrder(base.Ins, base.Opts)
+
+	if orderErr != nil {
+		return orderErr
+	}
+
+	log.Infof("Order placed for %s ID %s", base.Ticker, orderRes.ID)
+
+	return nil
 }