+++ release-notes.md
@@ -1,5 +1,17 @@
 # Mew Release Notes
 
+## [v0.1.2]
+2020-05-14
+
+* batch transactions
+    * Example command:
+    mew -lb -t @qqq_msft -v 1000
+    
+* preview
+    * Please confirm the order details.
+Order type: Market Sell Security: QQQ   Quantity: 0     price: 221.78 [y/n]
+
+
 ## [v0.1.1]
 2020-05-06